Use avi-lbaas endpoint for Avi Controller
In Avi Heat Plugin, we identify the Avi Controller IP from (a) avi-lbaas endpoint or (b) user sets the avi_controller field in heat.conf.
In our tests, we are testing with heat.conf. However, there are customers using avi-lbaas endpoint to pass the Avi Controller IP, we should test it.

There was a recent fix in this area: https://github.com/avinetworks/avi-heat/pull/44
Having a test to make sure the fix works fine.